requirements.txt
numpy
pandas
segmentation-models-pytorch
pytorch-lightning